Move the `getRenameEdits` function from the `LSPTask` to the `AbstractRenamer`
