+++ main/lsp/LSPTask.cc
@@ -247,37 +247,6 @@ bool LSPRequestTask::cancel(const MessageId &id) {
     return false;
 }
 
-void LSPTask::getRenameEdits(LSPTypecheckerInterface &typechecker, shared_ptr<AbstractRenamer> renamer,
-                             core::SymbolRef symbol, string newName) {
-    const core::GlobalState &gs = typechecker.state();
-    auto originalName = symbol.name(gs).show(gs);
-
-    renamer->addSymbol(symbol);
-
-    auto symbolQueue = renamer->getQueue();
-    for (auto sym = symbolQueue->pop(); sym.exists(); sym = symbolQueue->pop()) {
-        auto queryResult = queryBySymbol(typechecker, sym);
-        if (queryResult.error) {
-            return;
-        }
-
-        // Filter for untyped files, and deduplicate responses by location.  We don't use extractLocations here because
-        // in some cases like sends, we need the SendResponse to be able to accurately find the method name in the
-        // expression.
-        for (auto &response : filterAndDedup(gs, queryResult.responses)) {
-            auto loc = response->getLoc();
-            if (loc.file().data(gs).isPayload()) {
-                // We don't support renaming things in payload files.
-                return;
-            }
-
-            // We may process the same send multiple times in case of union types, but this is ok because the renamer
-            // de-duplicates edits at the same location
-            renamer->rename(response, sym);
-        }
-    }
-}
-
 bool LSPTask::needsMultithreading(const LSPIndexer &indexer) const {
     return false;
 }
