+++ main/lsp/requests/rename.cc
@@ -246,20 +246,20 @@ unique_ptr<ResponseMessage> RenameTask::runRequest(LSPTypecheckerInterface &type
         if (isValidRenameLocation(constResp->symbol, gs, response)) {
             auto originalName = constResp->symbol.name(gs).show(gs);
             renamer = makeRenamer(gs, config, constResp->symbol, params->newName);
-            getRenameEdits(typechecker, renamer, constResp->symbol, params->newName);
+            renamer->getRenameEdits(typechecker, constResp->symbol, params->newName);
             enrichResponse(response, renamer);
         }
     } else if (auto defResp = resp->isMethodDef()) {
         if (isValidRenameLocation(defResp->symbol, gs, response)) {
             renamer = makeRenamer(gs, config, defResp->symbol, params->newName);
-            getRenameEdits(typechecker, renamer, defResp->symbol, params->newName);
+            renamer->getRenameEdits(typechecker, defResp->symbol, params->newName);
             enrichResponse(response, renamer);
         }
     } else if (auto sendResp = resp->isSend()) {
         // We don't need to handle dispatchResult->secondary here, because it will be checked in getRenameEdits.
         auto method = sendResp->dispatchResult->main.method;
         renamer = makeRenamer(gs, config, method, params->newName);
-        getRenameEdits(typechecker, renamer, method, params->newName);
+        renamer->getRenameEdits(typechecker, method, params->newName);
         enrichResponse(response, renamer);
     }
